requirements.txt
Flask~=2.2.2
bcrypt~=4.0.0
PyJWT~=2.4.0
psycopg2-binary~=2.9.3